Move TransactionStatus to session package
